Code as of 1-26-17
Sorry, forgot to commit at the comps. anyways this is the code going
into the competition tomorrow, 1-27. There are some jerking-oscillating
issued in the PID witht he right side of the DR4B But that could be
partially caused by the chain bar. I will investigate at the competition
tomorrow. I'm pretty sure I can patch it in Ki.

As for other changes since december, Autons have been added, one base
auton for each color and a universal pylon auton.

Code was also changed to accomodate the recent switch to turbo motors on
the chain bar (Fun!). current solution kind of "throws" the cones onto
the stack; Not the <most> consistent, but incedibly fast. Seems good
enough as long as the cone was grasped properly by the rubber band
rollers.

Also lol I had no Idea I was competing tomorrow until about 10:00, and
it is 11:00 now.