Code as of the state championship on 3-3-18
Oh well. I'll hopefully qualify next year.

Anyways this includes code that is updated for the switch to a
front/back roller vs the previous side rollers. The autostacking code
was reverted back to the more consistent version. One of the autons was
extended for skills.

Oh and a few other things were changed since last update. I took a motor
off the mobile goal lift and put it on the drive alopng with our 12th
motor so we now have a 6 motor drive. Being able to push people was fun
lol.

As for the mobile goal lift, that was changed to the four bar version
for speed and the ability to place stacks in the 20 point zone, though
this didnt take many code changes.

Well, hopefully I'll do well at the National championship!