Code as of 4-7-18 NATIONAL CHAMPIONSHIP
GG, Fluffles the Exterminating Giraffe, you did well.

Anyways, for my final commit this season ( :( ) I include autonomouses
to grab both 2 and 3 cones, put them on a base, and put that base in the
20 point zone. Interestingly, and contrary to what I expected, a three
cone autonomous was the exception rather than the rule. Heck half of the
very best teams only had two cone autons.

Anyways also included are some fixes, such as to a mistake in the match
load function that would prevent the system from finishing resetting.

Also included are modifications to the PID system to accomodate the
addition of stops to the top end of the chain bar's movement.

Anyways, made it to QFs, got the highest combined scoring match in the
world for a few minutes in the last match of the season. I'll consider
that pretty good.

And as it seems that this is likely my last commit to the old Vex Cortex
system, I say it has served us well, even with its faults considered.